(Tech) move useIsCurrentUserEmail out of ui/components
